requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Security==1.7.4
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
WTForms==2.0.1
Werkzeug==0.9.6
backports.ssl-match-hostname==3.4.0.2
blinker==1.3
docker-py==0.6.0
itsdangerous==0.24
manage.py==0.2.6
passlib==1.6.2
requests==2.4.3
six==1.8.0
websocket-client==0.21.0
wsgiref==0.1.2
